fix split component
