--- components/Layout.tsx
@@ -1,19 +1,0 @@
-import { FC, ReactNode } from 'react'
-import { Header } from '@/components/Header'
-import { Footer } from '@/components/Footer'
-import { Sidebar } from '@/components/Sidebar'
-
-interface Children {
-  children: ReactNode
-}
-
-export const Layout: FC<Children> = ({ children }) => {
-  return (
-    <>
-      <Header />
-      <Sidebar />
-      <main>{children}</main>
-      <Footer />
-    </>
-  )
-}
